Remove standalone body-shape command

Copilot-Session: 43dbbaf0-9cf3-43c6-81c4-5f1b68de2738
