+++ docs/design/member-body-substrate.md
@@ -482,15 +482,15 @@ decompiles the selected API-surface bodies, and returns one row per match:
 stable source-facing member identity, MethodDef token, exact `PrintedExtent`,
 and the text selected by that extent. Bodies below `Full` fidelity, including
 state-machine kickoffs whose source body was not reconstructed, are explicit
-failures rather than searchable compiler plumbing. The `body-shape` CLI exposes
-that narrow query for one `--library`, acquires the portable PDB used by its
-source view, and passes its resolved user-facing `PrinterOptions`, so each
-extent indexes the same rendering. Library callers retain the low-level printer
-defaults and symbol source represented by their `MetadataSource` unless they
-supply options and a PDB explicitly. The command does not add a predicate
-language, persist an index, or infer parentage from IR object references. Public
-members are the default search boundary, `--all` expands that boundary, and
-`--limit` supplies explicit backpressure. `Instruction` belongs to the IL plane;
+failures rather than searchable compiler plumbing. The CLI exposes that query
+through the library, exact-type, and exact-member `Body Shapes`
+sections. Those hosts acquire the portable PDB used by their source view and
+pass resolved user-facing `PrinterOptions`, so each extent indexes the same
+rendering. Library callers retain the low-level printer defaults and symbol
+source represented by their `MetadataSource` unless they supply options and a
+PDB explicitly. The query does not persist an index or infer parentage from IR
+object references. Public members are the default search boundary and `--all`
+expands it. `Instruction` belongs to the IL plane;
 region-only and structural catalog names such as `CatchClause`, `SwitchSection`,
 and `Block` are likewise not exact body-shape query kinds.
