+++ src/dotnet-inspect.Tests/BodyShapeSearchTests.cs
@@ -1,18 +1,14 @@
 using System.Reflection;
-using System.Text.Json;
 using ILInspector.Decompiler;
 using ILInspector.Decompiler.Pipeline;
 using ILInspector.Metadata;
 using ILInspector.MetadataPrimitives;
-using DotnetInspector.Commands;
 using DotnetInspector.Fixtures;
-using DotnetInspector.Options;
-using DotnetInspector.Output;
 
 namespace DotnetInspector.Tests;
 
 [Collection("Console")]
-public sealed class BodyShapeCommandTests
+public sealed class BodyShapeSearchTests
 {
     static string FixturePath => typeof(BodyShapeFixture).Assembly.Location;
 
@@ -308,227 +304,5 @@ public sealed class BodyShapeCommandTests
             $"{genericMatch.TypeName}.",
             genericMatch.Member,
             StringComparison.Ordinal);
-    }
-
-    [Fact]
-    public async Task Command_TsvReportsExactCoordinatesAndPublicMatch()
-    {
-        var (exit, output, error) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "ObjectCreationExpression",
-                LibraryPath = FixturePath,
-                Tabular = true,
-                Tsv = true
-            })));
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Warning: Body-shape search skipped", error);
-        Assert.Contains("start_line\tstart_column\tend_line\tend_column", output);
-        Assert.Contains(nameof(BodyShapeFixture.PublicCreation), output);
-        Assert.DoesNotContain("PrivateCreation", output);
-        Assert.Contains("new object()", output);
-    }
-
-    [Fact]
-    public async Task Command_UsesExternalPdbForLocalNames()
-    {
-        string directory = Path.Combine(
-            Path.GetTempPath(),
-            $"body-shape-external-pdb-{Guid.NewGuid():N}");
-        Directory.CreateDirectory(directory);
-        string assemblyPath = Path.Combine(directory, Path.GetFileName(FixturePath));
-        File.Copy(FixturePath, assemblyPath);
-        try
-        {
-            var (exit, output, _) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-                BodyShapeCommand.Execute(new BodyShapeOptions
-                {
-                    Kind = "NameExpression",
-                    LibraryPath = assemblyPath,
-                    PdbPath = Path.ChangeExtension(FixturePath, ".pdb"),
-                    RenderOptions = StyleOptionCatalog.DefaultOptions,
-                    Tabular = true,
-                    Tsv = true
-                })));
-
-            Assert.Equal(0, exit);
-            var rows = output.Split('\n')
-                .Where(row => row.Contains(
-                    $".{nameof(BodyShapeFixture.ReadableLocal)}~",
-                    StringComparison.Ordinal))
-                .ToList();
-            Assert.Contains(rows, row => row.EndsWith("\tbuilder", StringComparison.Ordinal));
-            Assert.DoesNotContain(rows, row => row.EndsWith("\tstringBuilder", StringComparison.Ordinal));
-        }
-        finally
-        {
-            File.Delete(assemblyPath);
-            Directory.Delete(directory);
-        }
-    }
-
-    [Fact]
-    public async Task PdbAcquisition_PropagatesCallerCancellation()
-    {
-        string directory = Path.Combine(
-            Path.GetTempPath(),
-            $"body-shape-pdb-cancellation-{Guid.NewGuid():N}");
-        Directory.CreateDirectory(directory);
-        string runtimeAssembly = typeof(object).Assembly.Location;
-        string assemblyPath = Path.Combine(directory, Path.GetFileName(runtimeAssembly));
-        File.Copy(runtimeAssembly, assemblyPath);
-        try
-        {
-            using var httpClient = new HttpClient();
-            using var cancellation = new CancellationTokenSource();
-            cancellation.Cancel();
-
-            await Assert.ThrowsAnyAsync<OperationCanceledException>(() =>
-                ApiCommand.TryAcquirePdbPathAsync(
-                    assemblyPath,
-                    new ApiOptions { AssemblyPath = assemblyPath },
-                    new VerboseLogger(enabled: false),
-                    httpClient,
-                    cancellation.Token));
-        }
-        finally
-        {
-            File.Delete(assemblyPath);
-            Directory.Delete(directory);
-        }
-    }
-
-    [Fact]
-    public async Task Command_MarkdownHonorsColumnProjection()
-    {
-        var (exit, output, error) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "ObjectCreationExpression",
-                LibraryPath = FixturePath,
-                Columns = ["Token"]
-            })));
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Warning: Body-shape search skipped", error);
-        Assert.Contains("| Token |", output);
-        Assert.DoesNotContain("Member", output);
-        Assert.DoesNotContain("Start Line", output);
-        Assert.DoesNotContain("new object()", output);
-    }
-
-    [Fact]
-    public async Task Command_TsvHonorsKindOnlyProjection()
-    {
-        var (exit, output, error) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "ObjectCreationExpression",
-                LibraryPath = FixturePath,
-                Tabular = true,
-                Tsv = true,
-                Columns = ["Kind"]
-            })));
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Warning: Body-shape search skipped", error);
-        Assert.Equal(
-            ["kind", "ObjectCreationExpression"],
-            output.Trim().Split('\n').Distinct(StringComparer.Ordinal).ToArray());
-    }
-
-    [Fact]
-    public async Task Command_JsonPreservesMultiLineMatch()
-    {
-        var (exit, output, error) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "IfStatement",
-                LibraryPath = FixturePath,
-                JsonOutput = true
-            })));
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Warning: Body-shape search skipped", error);
-        using var document = JsonDocument.Parse(output);
-        var match = document.RootElement.EnumerateArray().Single(element =>
-            element.GetProperty("type_name").GetString() == typeof(BodyShapeFixture).FullName
-            && element.GetProperty("method_name").GetString() == nameof(BodyShapeFixture.Branch));
-        Assert.Equal(
-            $"0x{typeof(BodyShapeFixture).GetMethod(nameof(BodyShapeFixture.Branch))!.MetadataToken:X8}",
-            match.GetProperty("method_token").GetString());
-        Assert.Contains('\n', match.GetProperty("text").GetString()!);
-        Assert.True(
-            match.GetProperty("extent").GetProperty("end_line").GetInt32()
-            > match.GetProperty("extent").GetProperty("start_line").GetInt32());
-    }
-
-    [Fact]
-    public async Task Command_CountReportsNoMatches_AndUnreadableLibraryFails()
-    {
-        var noMatches = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "FixedStatement",
-                LibraryPath = FixturePath,
-                Count = true
-            })));
-
-        Assert.Equal(0, noMatches.ExitCode);
-        Assert.Equal("0", noMatches.Output.Trim());
-
-        string missing = Path.Combine(Path.GetTempPath(), $"missing-body-shape-{Guid.NewGuid():N}.dll");
-        var unreadable = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "ObjectCreationExpression",
-                LibraryPath = missing
-            })));
-
-        Assert.Equal(1, unreadable.ExitCode);
-            Assert.Equal(
-                $"Error: Could not read library: {missing}",
-                unreadable.Error.Trim());
-
-            string invalid = Path.Combine(Path.GetTempPath(), $"invalid-body-shape-{Guid.NewGuid():N}.dll");
-            await File.WriteAllTextAsync(
-                invalid,
-                "not a PE image",
-                TestContext.Current.CancellationToken);
-            try
-            {
-                var nonPe = await ConsoleCapture.RunAsync(() => Task.FromResult(
-                    BodyShapeCommand.Execute(new BodyShapeOptions
-                    {
-                        Kind = "ObjectCreationExpression",
-                        LibraryPath = invalid
-                    })));
-
-                Assert.Equal(1, nonPe.ExitCode);
-                Assert.Equal(
-                    $"Error: Could not read library: {invalid}",
-                    nonPe.Error.Trim());
-            }
-            finally
-            {
-                File.Delete(invalid);
-            }
-    }
-
-    [Fact]
-    public async Task Command_CountAppliesTheRenderedRowWindow()
-    {
-        var (exit, output, _) = await ConsoleCapture.RunAsync(() => Task.FromResult(
-            BodyShapeCommand.Execute(new BodyShapeOptions
-            {
-                Kind = "ObjectCreationExpression",
-                LibraryPath = FixturePath,
-                Rows = RowWindow.Range(2, 3),
-                Count = true
-            })));
-
-        Assert.Equal(0, exit);
-        Assert.Equal("2", output.Trim());
     }
 }
